create notification helper
